+++ duration/src/mod.ts
@@ -5,6 +5,7 @@
 export * as Time from './consts.ts';
 export * as Duration from './duration/mod.ts';
 export * as Format from './format.ts';
-export { humanize } from './duration/humanize.ts';
+export { humanize, type HumanizeOptions } from './duration/humanize.ts';
+export { getTranslations, supportedLocales, type HumanizeTranslations } from './duration/i18n.ts';
 export * from './time-types.ts';
 export { compareFields, duration, isEpochMilliseconds, isEpochSeconds, isField, isMilliseconds } from './utils.ts';